Add ui setting to toggle between IWD homepage and Default homepage. Began updating the loan channel query map to support new iwd loan channels.
