+++ src/pages/Homepage/DefaultHomepage.vue
@@ -1,0 +1,68 @@
+<template>
+	<div>
+		<hero-slideshow v-if="showSlideShow"
+			:promo-enabled="promoEnabled" :promo-content="promoContent"
+		/>
+		<monthly-good-explained v-if="promoEnabled" />
+		<why-kiva />
+		<category-grid />
+	</div>
+</template>
+
+<script>
+import _get from 'lodash/get';
+import contentfulCMS from '@/graphql/query/contentfulCMS.graphql';
+import { settingEnabled } from '@/util/settingsUtils';
+import WhyKiva from '@/components/Homepage/WhyKiva';
+import HeroSlideshow from './HeroSlideshow';
+import MonthlyGoodExplained from '@/components/Homepage/MonthlyGoodExplained';
+import CategoryGrid from '@/components/Homepage/CategoryGrid';
+import { processContent } from '@/util/contentfulUtils';
+
+export default {
+	components: {
+		WhyKiva,
+		HeroSlideshow,
+		MonthlyGoodExplained,
+		CategoryGrid,
+	},
+	data() {
+		return {
+			promoEnabled: false,
+			showSlideShow: null,
+			promoContent: {}
+		};
+	},
+	inject: ['apollo'],
+	mounted() {
+		this.apollo.query({
+			query: contentfulCMS,
+			variables: {
+				contentType: 'uiSetting',
+				contentKey: 'ui-homepage-promo',
+			}
+		}).then(({ data }) => {
+			// returns the contentful content of the uiSetting key ui-homepage-promo or empty object
+			// it should always be the first and only item in the array, since we pass the variable to the query above
+			const uiPromoSetting = _get(data, 'contentfulCMS.items', []).find(item => item.key === 'ui-homepage-promo'); // eslint-disable-line max-len
+			this.promoEnabled = settingEnabled(
+				uiPromoSetting,
+				'active',
+				'startDate',
+				'endDate'
+			);
+			this.promoContent = processContent(uiPromoSetting.content);
+		}).finally(() => {
+			this.showSlideShow = true;
+		});
+	},
+};
+</script>
+
+<style lang="scss" scoped>
+@import 'settings';
+
+.page-content {
+	padding: 1.625rem 0;
+}
+</style>
